% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbmim.R
\name{cbmim}
\alias{cbmim}
\title{Fit CKMR/CMIM}
\usage{
cbmim(
  y,
  x,
  z,
  niter = 10000,
  nburn = 0.5 * niter,
  nthin = 10,
  nchains = 1,
  ncores = 1,
  hierarchical = FALSE,
  polar = FALSE,
  prior_theta_kappa = 1,
  prior_pi_additive = c(1, 1),
  prior_tau2_additive = c(1, 1),
  prior_pi_nonadditive = c(1, 1),
  prior_rho_nonadditive = c(1, 10),
  prior_tau2_nonadditive = c(1, 1),
  prior_sigma2 = c(1, 1),
  stepsize_theta_kappa = 100,
  stepsize_tau2_nonadditive = 0.1,
  stepsize_rho_nonadditive = 0.1,
  oversample = FALSE,
  prop_phi_a = 200,
  invmethod = c("exact"),
  rank = 100,
  knots = NULL,
  approxProj = FALSE,
  kernelfun = "gaussian",
  basis.opts.list = NULL,
  draw_h = FALSE,
  centering = TRUE,
  scaling = TRUE,
  startvals = list()
)
}
\arguments{
\item{y}{response}

\item{x}{list of exposures/mixture components}

\item{z}{confounders to be adjusted}

\item{niter}{number of iterations}

\item{nburn}{burn-in fraction}

\item{nthin}{thinning number}

\item{nchains}{number of chains ## not yet implemented}

\item{ncores}{number of cores for mclapply (set to 1 for non-parallel) ## only used if nchains>1}

\item{hierarchical}{for BMIM, set to TRUE for hierarchically well formulated structure}

\item{polar}{set to TRUE for polar coordinate parameterization}

\item{prior_theta_kappa}{kappa hyperparameter for vMF prior (only used in bmim) (Antoniadis used 100-700)}

\item{prior_pi_additive}{prior for pi in spike & slab on additive component}

\item{prior_tau2_additive}{shape and rate for inverse gamma on spline penalty terms}

\item{prior_pi_nonadditive}{prior for pi in spike & slab on non-additive component}

\item{prior_rho_nonadditive}{shape and rate for gamma}

\item{prior_tau2_nonadditive}{shape and rate for inverse gamma on nonadditive GP component}

\item{prior_sigma2}{shape and rate for inverse gamma on sigma^2}

\item{stepsize_theta_kappa}{kappa in vMF proposal for theta (bmim only)  (Antoniadis used 1000)}

\item{stepsize_tau2_nonadditive}{jumpsize/sd for gamma proposal on tau2_nonadditive}

\item{stepsize_rho_nonadditive}{jumpsize/sd for gamma proposal on rho}

\item{oversample}{oversample the between model moves}

\item{prop_phi_a}{proposal step size for phi}

\item{invmethod}{"exact" for full matrix inversion, "GPP" for GP projection, "rpSMW" for random projection +SMW, "lzSMWmax/min/both" is lanczos approximation with largest/smallest/both m eigencomponents +SMW, "lzDIRECTmax/min/both" is lanczos approximation to full matrix (I+PKP)}

\item{rank}{rank of approximation for lz/rp approximations of PKP (SMW) or I+PKP (DIRECT) or number of knots for GPP approach}

\item{knots}{optional set of knots for GPP (overrides rank)}

\item{approxProj}{set to TRUE to use the approximate projection matrix (i.e. dont update P as theta changes)}

\item{kernelfun}{choice of kernel function}

\item{basis.opts.list}{choice of basis representations (for bkmr-DLM) #list(type="face", pve=.9)}

\item{draw_h}{draw h or not}

\item{centering}{should covariates be centered to have mean=0}

\item{scaling}{should covariates be scaled to have SD=1}

\item{startvals}{pass any chosen starting values}
}
\value{
Returns posterior samples
}
\description{
Fit collapsed kernel machine regression or multiple index model.
}
