% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lastvals.R
\name{get_lastvals}
\alias{get_lastvals}
\title{Get last values.}
\usage{
get_lastvals(fit)
}
\arguments{
\item{fit}{Fitted model}
}
\value{
Returns list of last values
}
\description{
Extract parameter values from last iteration.
}
