% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PIPs.R
\name{get_PIPs}
\alias{get_PIPs}
\title{Get PIPs}
\usage{
get_PIPs(fit)
}
\arguments{
\item{fit}{Fitted model}
}
\value{
List of PIP objects.
}
\description{
Get main effect and second order interaction posterior inclusion probabilities.
}
