% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_params.R
\name{initialize_params}
\alias{initialize_params}
\title{Initialize parameters}
\usage{
initialize_params(
  y,
  x,
  z,
  prior_theta_kappa,
  prior_pi_additive,
  prior_tau2_additive,
  prior_pi_nonadditive,
  prior_rho_nonadditive,
  prior_tau2_nonadditive,
  prior_sigma2,
  stepsize_theta_kappa,
  stepsize_tau2_nonadditive,
  stepsize_rho_nonadditive,
  oversample,
  prop_phi_a,
  kernelfun,
  invmethod,
  rank,
  knots,
  approxProj,
  hierarchical,
  polar,
  startvals = list(),
  basis.opts.list
)
}
\description{
Initialize parameters
}
\keyword{internal}
