% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_surface.R
\name{pred_surface}
\alias{pred_surface}
\title{Predict surface}
\usage{
pred_surface(fit, Xnew = NULL, gridlen = 21, includeInt = TRUE)
}
\arguments{
\item{fit}{Fitted model}

\item{Xnew}{New exposure values for manual/arbitrary prediction}

\item{gridlen}{Length of exposure grid}

\item{includeInt}{Include intercept in prediction}
}
\value{
Returns list of predictions
}
\description{
Predict regression surface, componentwise by default, holding other exposures at 0
}
