% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_surface.R
\name{pred_surface_indexwise}
\alias{pred_surface_indexwise}
\title{Predict surface}
\usage{
pred_surface_indexwise(fit, gridlen = 21, includeInt = TRUE)
}
\arguments{
\item{fit}{Fitted model}

\item{gridlen}{Length of exposure grid}

\item{includeInt}{Include intercept in prediction}
}
\value{
Returns list of predictions
}
\description{
Predict regression surface, indexwise, holding other indices at 0
}
