% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_surface.R
\name{pred_surface_indexwise_interac}
\alias{pred_surface_indexwise_interac}
\title{Indexwise predictions with interactions}
\usage{
pred_surface_indexwise_interac(
  fit,
  gridlen = 21,
  includeInt = TRUE,
  restrict = TRUE
)
}
\arguments{
\item{fit}{Fitted model}

\item{gridlen}{Length of exposure grid}

\item{includeInt}{Include intercept in prediction}

\item{restrict}{restrict to non-null associations (only for interactions)}
}
\value{
Returns list of predictions
}
\description{
Indexwise predictions with interactions
}
