% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_CKMR}
\alias{run_CKMR}
\title{Run CKMR Experiment}
\usage{
run_CKMR(
  s1_,
  s2_,
  s3_,
  s4_,
  numReps_,
  numThreads_,
  networkParameters_,
  reproductionReference_,
  patchReleases_,
  migrationMale_,
  migrationFemale_,
  migrationBatch_,
  sampDays_,
  sampCov_,
  outputDirectory_,
  verbose_
)
}
\description{
Run multiple CKMR experiments.
}
\examples{
\dontrun{
nope
none
sorry
}

}
