% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_surface.R
\name{summarize_pred}
\alias{summarize_pred}
\title{Sumarize prediction intervals}
\usage{
summarize_pred(pred, pct = 0.95, assoc = TRUE, centered = TRUE)
}
\arguments{
\item{pred}{Predictions}

\item{pct}{Pct level for CIs}

\item{assoc}{Estimate contrast (T/F)}

\item{centered}{Center estimates (T/F) only used if assoc=FALSE}
}
\value{
Returns list of predictions
}
\description{
Summarize prediction intervals results of a call to pred_surface
}
