% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_surface.R
\name{summarize_pred_mat}
\alias{summarize_pred_mat}
\title{Summarize prediction intervals for single matrix ofresults to be called by summarize_pred().}
\usage{
summarize_pred_mat(pred, pct = 0.95)
}
\description{
Summarize prediction intervals for single matrix ofresults to be called by summarize_pred().
}
\keyword{internal}
