% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_surface.R
\name{summarize_pred_PSR}
\alias{summarize_pred_PSR}
\title{Summarize PSR for a call to pred_surface}
\usage{
summarize_pred_PSR(pred, nchains)
}
\description{
Summarize PSR for a call to pred_surface
}
\keyword{internal}
