% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PIPs.R
\name{summarize_weights}
\alias{summarize_weights}
\title{Summarize weights}
\usage{
summarize_weights(fit)
}
\arguments{
\item{fit}{Fitted model}
}
\value{
List of theta and w summaries
}
\description{
Summarize thetas and ws
}
