% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_index.R
\name{update_gamma_index}
\alias{update_gamma_index}
\title{Update gamma_index}
\usage{
update_gamma_index(params)
}
\description{
Between model moves for bmim (gamma_additive, beta, gamma_nonadditive, rho, theta)
}
\keyword{internal}
