% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hierarchical.R
\name{update_gamma_index_hierarchical}
\alias{update_gamma_index_hierarchical}
\title{Update gamma_index}
\usage{
update_gamma_index_hierarchical(params)
}
\description{
Update gamma_index
}
\keyword{internal}
