% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hierarchical.R
\name{update_gamma_index_hierarchical_polar}
\alias{update_gamma_index_hierarchical_polar}
\title{Update gamma_index_hierarchical (Polar)}
\usage{
update_gamma_index_hierarchical_polar(params)
}
\description{
Update gamma_index_hierarchical (Polar)
}
\keyword{internal}
