% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_nonadditive.R
\name{update_gamma_nonadditive}
\alias{update_gamma_nonadditive}
\title{Update gamma nonadditive and rho}
\usage{
update_gamma_nonadditive(params)
}
\description{
Update gamma nonadditive and rho
}
\keyword{internal}
