% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hierarchical.R
\name{update_pi_nonadditive_hierarchical}
\alias{update_pi_nonadditive_hierarchical}
\title{Update pi_nonadditive}
\usage{
update_pi_nonadditive_hierarchical(params)
}
\description{
Update pi_nonadditive
}
\keyword{internal}
