% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_nonadditive.R
\name{update_rho_nonadditive}
\alias{update_rho_nonadditive}
\title{Refinement step for rho_nonadditive}
\usage{
update_rho_nonadditive(params)
}
\description{
Refinement step for rho_nonadditive
}
\keyword{internal}
